--- a/.gitignore
+++ b/.gitignore
@@ -14,3 +14,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
+coverage/*