--- a/package.json
+++ b/package.json
@@ -47,10 +47,10 @@
     "sinon": "^10.0.0"
   },
   "dependencies": {
+    "@stablelib/chacha20poly1305": "^1.0.1",
     "@stablelib/hkdf": "^1.0.1",
     "@stablelib/sha256": "^1.0.1",
     "@stablelib/x25519": "^1.0.1",
-    "aead-js": "^0.1.0",
     "debug": "^4.3.1",
     "it-buffer": "^0.1.1",
     "it-length-prefixed": "^5.0.2",