Replaced:
 - bcrypto/x25519 with  https://github.com/StableLib/stablelib/tree/master/packages/x25519
 - bcrypto/aead with https://github.com/StableLib/stablelib/tree/master/packages/chacha20poly1305
 - bcrypto/hkdf with https://github.com/StableLib/stablelib/tree/master/packages/hkdf

resolves #93