@vasco-santos 2021-06-11
🎉 

Do you have any metrics on the bundle size difference?