@vasco-santos 2021-06-11
~27% decrease on its bundle size, from 191.5kb to 141.2kb (minified+gzipped)

🚀 ❤️