@mpetrunic 2021-06-11
> ~27% decrease on its bundle size, from 191.5kb to 141.2kb (minified+gzipped)
> 
> 🚀 ❤️

from our benchmark there is also speed improvment:
```
NEW:

Initializing handshake benchmark
Init complete, running benchmark
handshake x 39.97 ops/sec ±4.98% (67 runs sampled)
Done in 6.04s.


OLD:

Initializing handshake benchmark
Init complete, running benchmark
handshake x 27.23 ops/sec ±4.57% (67 runs sampled)
Done in 6.10s.
```